Add files via upload
The conceptual nature of this proposal embodies how an open society benefits from all of its members being allowed to participate in its well-being. ‘The whole can be greater than any one of its parts’, an old but true cliché. This is a core tenet of Open Source values. 
We are inviting the public to participate in an art process via the web (https://vsoch.github.io/opensource-art/about/), as well as allowing me as an artist to express the strength of inclusiveness with several paintings. These paintings will represent how we are all connected, and how we need a society that does not build walls, but rather facilitates strength by inclusion. 
The work of the public contributors will be visible by opening the URL provided. People may view it on their phones, their tablets, or their computers. 
In addition, the project itself if being done by collaboration between artist and software engineer using Github as the medium in its entire collaborative process. GitHub as well as RedHat embodies open source values.