probably last sentence needs to be elsewhere
Natacha Sochat and Vanessa Sochat timeline.

At present the all the paintings that will be displayed are not fully completed, but will be 5’x6’ to 6’x6’ in size, three to four large paintings on the lobby wall. The code for the exhibition is being worked on currently as well. All will be completed by time of exhibition date.
This will be a fluid project proposal as there are varied possibilities for display. I will supply rough sketches of each of these ideas and will work with RedHat on the final decision: There will be a URL posted in the lobby where people can go and participate.
1.	Two large paintings which will be connected by crocheting to a number of smaller works between them. These smaller works may embody some of the uploaded imagery translated into a painting or perhaps simply a graphic copy of the uploaded work. These smaller works will be added each week as the project evolves during the exhibition period at RedHat. It is possible that I may be able to crochet the connection while in the lobby as a performative element each week for the installation of the additional smaller works.
2.	Three large paintings with one large painting in the middle which will represent a tapestry/quilt format of imagery metaphorically showing how many distinct parts can create a beautiful whole.
3.	Two large paintings connected to smaller ‘containers’ (small paintings/objects) representing collaboration and connectedness. As in #1, the connection will be crocheting.
I am using crocheting as a connector because for me as an artist it involves the hand and my labor. The hand is the mind, and labor is a valuable social force that always needs to be valued.